requirements.txt
Django==4.2.1
djangorestframework
djangorestframework_simplejwt